@@ -26,38 +26,40 @@
     "babel-preset-es2015": "^6.9.0",
     "babel-preset-react": "^6.11.1",
     "babel-preset-stage-0": "^6.5.0",
-    "babel-relay-plugin": "^0.9.3",
-    "babel-relay-plugin-0.9": "^0.9.4",
-    "babel-relay-plugin-loader-0.9": "^0.9.6",
+    "babel-relay-plugin": "^0.10.0",
+    "babel-relay-plugin-loader": "gauravtiwari/babel-relay-plugin-loader",
     "classnames": "^2.2.5",
     "css-loader": "^0.23.1",
-    "dialog-polyfill": "^0.4.4",
-    "dompurify": "^0.8.3",
+    "dialog-polyfill": "^0.4.5",
+    "dompurify": "^0.8.4",
     "expose-loader": "^0.7.1",
     "extract-text-webpack-plugin": "^1.0.1",
-    "formsy-material-ui": "0.5.2",
-    "formsy-react": "^0.18.1",
-    "highlight.js": "^9.8.0",
+    "formsy-material-ui": "^0.5.3",
+    "formsy-react": "^0.19.0",
+    "highlight.js": "^9.9.0",
     "immutability-helper": "^2.0.0",
     "imports-loader": "^0.6.5",
-    "jquery": "^3.1.0",
+    "jquery": "^3.1.1",
     "jquery-ujs": "^1.2.2",
     "json-loader": "^0.5.4",
-    "material-ui": "^0.16.4",
-    "moment": "^2.14.1",
-    "node-sass": "^3.12.5",
+    "marked": "^0.3.6",
+    "material-ui": "^0.16.5",
+    "moment": "^2.17.1",
+    "node-sass": "^4.0.0",
     "promise-window": "^1.1.0",
     "query-string": "^4.2.3",
-    "react": "^15.4.0",
-    "react-dom": "^15.4.0",
+    "react": "^15.4.1",
+    "react-dom": "^15.4.1",
+    "react-dropzone": "^3.7.3",
     "react-loader": "^2.4.0",
-    "react-relay": "^0.9.3",
+    "react-relay": "^0.10.0",
+    "react-rte": "^0.11.0",
     "react-tap-event-plugin": "^2.0.1",
-    "react-truncate": "^2.0.2",
+    "react-truncate": "^2.0.3",
     "sass-loader": "^4.0.2",
     "style-loader": "^0.13.1",
     "underscore": "^1.8.3",
-    "webpack": "^1.13.2"
+    "webpack": "^1.14.0"
   },
   "license": "MIT",
   "engines": {
