@@ -5,15 +5,28 @@ import { Card, CardActions, CardTitle, CardText } from 'material-ui/Card';
 import moment from 'moment';
 import IconButton from 'material-ui/IconButton';
 import FontIcon from 'material-ui/FontIcon';
-
+import marked from 'marked';
 import Languages from '../../../utils/languages.json';
 
 // Child Components icons
 import GithubIcon from '../../shared/icons/github.es6';
-import sanitize from '../../../utils/sanitize.es6';
+import { sanitizeText } from '../../../utils/sanitize.es6';
+
+marked.setOptions({
+  renderer: new marked.Renderer(),
+  gfm: true,
+  tables: true,
+  breaks: false,
+  pedantic: false,
+  sanitize: true,
+  smartLists: true,
+  smartypants: false,
+});
 
 const Github = (props) => {
   const { achievement, remove } = props;
+  const starFields = { pr: 'comments', repo: 'stargazers_count' };
+  const count = achievement[starFields[achievement.category]];
 
   return (
     <div className={`achievement ${achievement.source_name}`}>
@@ -27,7 +40,7 @@ const Github = (props) => {
             <div className="achievement-card-content">
               <h2 className="intro">
                 <i className="icon material-icons">lock_open</i>
-                <span>Open Source </span>
+                <span>Open Source</span>
               </h2>
 
               {achievement.is_owner ?
@@ -58,7 +71,7 @@ const Github = (props) => {
                       target="_blank"
                       rel="noopener noreferrer"
                     >
-                      {achievement.name.replace(/[_-]/g, ' ')}
+                      {achievement.name || achievement.title}
                     </a>
                   </div>
                 }
@@ -67,22 +80,28 @@ const Github = (props) => {
               <CardText
                 className="achievement-card-description"
                 dangerouslySetInnerHTML={{
-                  __html: sanitize(achievement.description),
+                  __html: sanitizeText(achievement.description || marked(achievement.body)),
                 }}
               />
 
               <CardActions className="meta">
-                <span
-                  className="badge"
-                  style={{
-                    backgroundColor: Languages[achievement.language],
-                  }}
-                >
-                  {achievement.language}
+                <span className="badge">
+                  {achievement.category}
                 </span>
 
+                {achievement.language ?
+                  <span
+                    className="badge"
+                    style={{
+                      backgroundColor: Languages[achievement.language],
+                    }}
+                  >
+                    {achievement.language}
+                  </span> : ''
+                }
+
                 <span className="badge">
-                  {`${achievement.stargazers_count}`}
+                  {`${count}`}
                   <FontIcon
                     color="#fff"
                     className="material-icons"
@@ -115,13 +134,17 @@ const GithubContainer = Relay.createContainer(Github, {
       fragment on Import {
         id,
         name,
+        title,
         source_name,
         description,
+        category,
+        body,
         developer_id,
         connection_id,
         language,
         html_url,
         is_owner,
+        comments,
         stargazers_count,
         pinned,
         created_at,
