@@ -1,25 +1,18 @@
 // Modules
 import React from 'react';
 import Relay from 'react-relay';
-import createDOMPurify from 'dompurify';
-
-// Stylesheets
+import { sanitizeText } from '../../utils/sanitize.es6';
 import '../styles/bio.sass';
 
-const Bio = (props) => {
-  const bio = createDOMPurify.sanitize(
-    props.developer.bio,
-    { ALLOWED_TAGS: ['b', 'i'] }
-  );
-
-  return (
-    <div className="bio">
-      {props.developer.bio ?
-        <span dangerouslySetInnerHTML={{ __html: bio }} /> : ''
-      }
-    </div>
-  );
-};
+const Bio = props => (
+  <div className="bio">
+    {props.developer.bio ?
+      <span
+        dangerouslySetInnerHTML={{ __html: sanitizeText(props.developer.bio) }}
+      /> : ''
+    }
+  </div>
+);
 
 Bio.propTypes = {
   developer: React.PropTypes.object,
