@@ -4,24 +4,29 @@ import Relay from 'react-relay';
 import { ListItem } from 'material-ui/List';
 import FontIcon from 'material-ui/FontIcon';
 import Checkbox from 'material-ui/Checkbox';
-import createDOMPurify from 'dompurify';
-
-// Stylesheet
+import { sanitizeText } from '../../../utils/sanitize.es6';
 import '../../styles/pins.sass';
 
 const Data = (props) => {
   const { item, toggleItemOnServer } = props;
-  const description = createDOMPurify.sanitize(
-    item.description || item.body || item.summary || item.tagline,
-    { ALLOWED_TAGS: ['b', 'i'] }
+  const description = sanitizeText(
+    item.description || item.body || item.summary || item.tagline
   );
 
-  const starFields = new Map();
-  starFields.set('github', 'stargazers_count');
-  starFields.set('stackoverflow', 'up_vote_count');
-  starFields.set('youtube', 'likeCount');
-  starFields.set('meetup', 'yes_rsvp_count');
-  starFields.set('producthunt', 'votes_count');
+  const starFields = {
+    github: {
+      pr: 'comments',
+      repo: 'stargazers_count',
+    },
+    stackoverflow: 'up_vote_count',
+    youtube: 'likeCount',
+    meetup: 'yes_rsvp_count',
+    producthunt: 'votes_count',
+  };
+
+  const count = item.source_name === 'github' ?
+    item[starFields[item.source_name][item.category]] :
+    item[starFields[item.source_name]];
 
   return (
     <ListItem
@@ -45,7 +50,7 @@ const Data = (props) => {
               color: '#777',
             }}
           >
-            {item[starFields.get(item.source_name)]}
+            {count}
             <FontIcon
               color="#777"
               className="material-icons"
@@ -85,6 +90,7 @@ const DataContainer = Relay.createContainer(Data, {
         name,
         body,
         description,
+        category,
         stargazers_count,
         likeCount,
         viewCount,
@@ -96,6 +102,7 @@ const DataContainer = Relay.createContainer(Data, {
         votes_count,
         comments_count,
         discussion_url,
+        comments,
         tagline,
         pinned,
       }
