@@ -2,7 +2,7 @@ import React from 'react';
 import Loader from 'react-loader';
 
 const LoadingComponent = props => (
-  <div className={`overlay ${props.cssClass}`}>
+  <div className={`overlay ${props.cssClass || 'absolute'}`}>
     <Loader loaded={false} />
   </div>
 );
