@@ -9,11 +9,10 @@ class Connection < ApplicationRecord
 
   belongs_to :developer, touch: true
   has_many :imports, dependent: :destroy
-
   after_commit :import_connection_data, if: :access_token_previously_changed?
 
-  validates_presence_of :provider
-  validates_uniqueness_of :provider
+  validates_presence_of :provider, :developer_id
+  validates_uniqueness_of :provider, scope: [:uid, :developer_id]
 
   def owner?(user)
     user == developer
