@@ -11,7 +11,7 @@ class Authenticator
 
   def find_or_create_from_oauth
     ActiveRecord::Base.transaction do
-      @connection = Connection.where(connection_attrs).first_or_create
+      @connection = Connection.where(connection_attrs).first_or_initialize
       @connection.developer = create_from_oauth if @connection.developer_id.nil?
       @connection.access_token = auth.credentials.token
       @connection.developer if @connection.save!
@@ -23,7 +23,7 @@ class Authenticator
   # rubocop:disable Metrics/MethodLength
   # rubocop:disable Metrics/AbcSize
   def create_from_oauth
-    Developer.new(
+    Developer.create(
       email: auth.info.email,
       bio: auth.extra.raw_info.bio,
       password: Devise.friendly_token[0, 20],
