@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20161120045241) do
+ActiveRecord::Schema.define(version: 20161201015109) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -26,7 +26,7 @@ ActiveRecord::Schema.define(version: 20161120045241) do
     t.datetime "created_at",                   null: false
     t.datetime "updated_at",                   null: false
     t.index ["developer_id"], name: "index_connections_on_developer_id", using: :btree
-    t.index ["uid", "provider"], name: "index_connections_on_uid_and_provider", unique: true, using: :btree
+    t.index ["uid", "provider", "developer_id"], name: "index_connections_on_uid_and_provider_and_developer_id", unique: true, using: :btree
   end
 
   create_table "developers", force: :cascade do |t|
@@ -136,13 +136,90 @@ ActiveRecord::Schema.define(version: 20161120045241) do
     t.integer  "developer_id"
     t.datetime "created_at",                    null: false
     t.datetime "updated_at",                    null: false
+    t.string   "category"
+    t.index ["category"], name: "index_imports_on_category", using: :btree
     t.index ["connection_id"], name: "index_imports_on_connection_id", using: :btree
     t.index ["data"], name: "index_imports_on_data", using: :gin
     t.index ["developer_id"], name: "index_imports_on_developer_id", using: :btree
     t.index ["pinned"], name: "achievements", where: "(pinned = true)", using: :btree
     t.index ["pinned"], name: "index_imports_on_pinned", using: :btree
     t.index ["source_id"], name: "index_imports_on_source_id", using: :btree
-    t.index ["source_name", "source_id", "connection_id"], name: "index_imports_on_source_name_and_source_id_and_connection_id", unique: true, using: :btree
+    t.index ["source_name", "source_id", "connection_id", "category"], name: "unique_import_per_category_and_source", unique: true, using: :btree
+  end
+
+  create_table "mailboxer_conversation_opt_outs", force: :cascade do |t|
+    t.string  "unsubscriber_type"
+    t.integer "unsubscriber_id"
+    t.integer "conversation_id"
+    t.index ["conversation_id"], name: "index_mailboxer_conversation_opt_outs_on_conversation_id", using: :btree
+    t.index ["unsubscriber_id", "unsubscriber_type"], name: "index_mailboxer_conversation_opt_outs_on_unsubscriber_id_type", using: :btree
+  end
+
+  create_table "mailboxer_conversations", force: :cascade do |t|
+    t.string   "subject",        default: ""
+    t.integer  "messages_count"
+    t.integer  "receipts_count"
+    t.datetime "created_at",                  null: false
+    t.datetime "updated_at",                  null: false
+  end
+
+  create_table "mailboxer_notifications", force: :cascade do |t|
+    t.string   "type"
+    t.text     "body"
+    t.string   "subject",              default: ""
+    t.string   "sender_type"
+    t.integer  "sender_id"
+    t.integer  "conversation_id"
+    t.integer  "receipts_count"
+    t.boolean  "draft",                default: false
+    t.string   "notification_code"
+    t.string   "notified_object_type"
+    t.integer  "notified_object_id"
+    t.string   "message_object_type"
+    t.integer  "message_object_id"
+    t.string   "attachment"
+    t.datetime "updated_at",                           null: false
+    t.datetime "created_at",                           null: false
+    t.boolean  "global",               default: false
+    t.datetime "expires"
+    t.index ["conversation_id"], name: "index_mailboxer_notifications_on_conversation_id", using: :btree
+    t.index ["message_object_type", "message_object_id"], name: "mailboxer_message_object", using: :btree
+    t.index ["notification_code"], name: "index_mailboxer_notifications_on_notification_code", using: :btree
+    t.index ["notified_object_id", "notified_object_type"], name: "index_mailboxer_notifications_on_notified_object_id_and_type", using: :btree
+    t.index ["notified_object_type", "notified_object_id"], name: "mailboxer_notified_object", using: :btree
+    t.index ["sender_id", "sender_type"], name: "index_mailboxer_notifications_on_sender_id_and_sender_type", using: :btree
+    t.index ["type"], name: "index_mailboxer_notifications_on_type", using: :btree
+  end
+
+  create_table "mailboxer_receipts", force: :cascade do |t|
+    t.string   "receiver_type"
+    t.integer  "receiver_id"
+    t.integer  "notification_id",                                null: false
+    t.boolean  "is_read",                        default: false
+    t.boolean  "trashed",                        default: false
+    t.boolean  "deleted",                        default: false
+    t.string   "mailbox_type",        limit: 25
+    t.integer  "messages_count"
+    t.integer  "notifications_count"
+    t.datetime "created_at",                                     null: false
+    t.datetime "updated_at",                                     null: false
+    t.boolean  "is_delivered",                   default: false
+    t.boolean  "is_rejected",                    default: false
+    t.boolean  "is_spam",                        default: false
+    t.boolean  "is_replied",                     default: false
+    t.string   "delivery_method"
+    t.string   "message_id"
+    t.index ["deleted"], name: "deleted_receipts", where: "(deleted = true)", using: :btree
+    t.index ["is_read"], name: "read_receipts", where: "(is_read = true)", using: :btree
+    t.index ["is_read"], name: "unread_receipts", where: "(is_read = false)", using: :btree
+    t.index ["mailbox_type", "trashed", "deleted"], name: "all_inbox_receipts", where: "(((mailbox_type)::text = 'inbox'::text) AND (trashed = false) AND (deleted = false))", using: :btree
+    t.index ["mailbox_type", "trashed", "deleted"], name: "all_sentbox_receipts", where: "(((mailbox_type)::text = 'sentbox'::text) AND (trashed = false) AND (deleted = false))", using: :btree
+    t.index ["mailbox_type"], name: "inbox_receipts", where: "((mailbox_type)::text = 'inbox'::text)", using: :btree
+    t.index ["mailbox_type"], name: "sent_receipts", where: "((mailbox_type)::text = 'sentbox'::text)", using: :btree
+    t.index ["mailbox_type"], name: "trashed_receipts", where: "((trashed = true) AND (deleted = false))", using: :btree
+    t.index ["notification_id"], name: "index_mailboxer_receipts_on_notification_id", using: :btree
+    t.index ["receiver_id", "receiver_type"], name: "index_mailboxer_receipts_on_receiver_id_and_receiver_type", using: :btree
+    t.index ["trashed"], name: "not_trashed_receipts", where: "(trashed = false)", using: :btree
   end
 
   create_table "que_jobs", primary_key: ["queue", "priority", "run_at", "job_id"], force: :cascade, comment: "3" do |t|
@@ -160,4 +237,7 @@ ActiveRecord::Schema.define(version: 20161120045241) do
   add_foreign_key "favourites", "employers"
   add_foreign_key "imports", "connections"
   add_foreign_key "imports", "developers"
+  add_foreign_key "mailboxer_conversation_opt_outs", "mailboxer_conversations", column: "conversation_id", name: "mb_opt_outs_on_conversations_id"
+  add_foreign_key "mailboxer_notifications", "mailboxer_conversations", column: "conversation_id", name: "notifications_on_conversation_id"
+  add_foreign_key "mailboxer_receipts", "mailboxer_notifications", column: "notification_id", name: "receipts_on_notification_id"
 end
